+++ src/View/Helper/FormHelper.php
@@ -15,34 +15,36 @@ declare(strict_types=1);
 namespace MeTools\View\Helper;
 
 use Cake\Utility\Hash;
-use Cake\View\Helper\FormHelper as CakeFormHelper;
+use Cake\View\Helper\FormHelper as BaseFormHelper;
 use Cake\View\View;
-use MeTools\View\Widget\HiddenWidget;
 
 /**
  * Provides functionalities for forms
  * @property \MeTools\View\Helper\IconHelper $Icon
  * @property \MeTools\View\Helper\HtmlHelper $Html
- * @deprecated 2.21.5 Use instead `BootstrapFormHelper`
+ * @property \Cake\View\Helper\UrlHelper $Url
  */
-class FormHelper extends CakeFormHelper
+class FormHelper extends BaseFormHelper
 {
     /**
      * Helpers
      * @var array
      */
     public $helpers = [
-        'Html' => ['className' => 'MeTools.Html'],
+        'MeTools.Html',
         'MeTools.Icon',
         'Url',
     ];
 
     /**
-     * Property to check if we're working with an inline form.
-     * It's changed by `createInline()` method.
      * @var bool
      */
-    protected bool $inline = false;
+    protected bool $isInline = false;
+
+    /**
+     * @var bool
+     */
+    protected bool $isPost;
 
     /**
      * Construct the widgets and binds the default context providers.
@@ -50,45 +52,109 @@ class FormHelper extends CakeFormHelper
      * This method only rewrites the default templates config.
      * @param \Cake\View\View $view The View this helper is being attached to
      * @param array $config Configuration settings for the helper
-     * @return void
      */
     public function __construct(View $view, array $config = [])
     {
-        deprecationWarning('`FormHelper` is deprecated. Use instead `BootstrapFormHelper`');
-
         //Rewrites default templates config
         $this->_defaultConfig = Hash::merge($this->_defaultConfig, ['templates' => [
-            'checkbox' => '<input class="form-check-input" type="checkbox" name="{{name}}" value="{{value}}"{{attrs}}>',
-            'checkboxContainer' => '<div class="form-check input {{type}}{{required}}">{{content}}</div>',
-            'nestingLabel' => '{{hidden}}<label{{attrs}}>{{input}} {{text}}{{help}}</label>',
-            'hidden' => '<input type="{{type}}" name="{{name}}"{{attrs}}/>',
-            'input' => '<input type="{{type}}" name="{{name}}" class="form-control"{{attrs}}/>',
-            'inputError' => '<input type="{{type}}" name="{{name}}" class="form-control is-invalid"{{attrs}}/>',
-            'inputContainer' => '<div class="form-group input {{type}}{{required}}">{{content}}{{help}}</div>',
-            'inputContainerError' => '<div class="form-group input {{type}}{{required}} error">{{content}}{{error}}</div>',
-            'select' => '<select name="{{name}}" class="form-control"{{attrs}}>{{content}}</select>',
-            'selectMultiple' => '<select name="{{name}}[]" multiple="multiple" class="form-control"{{attrs}}>{{content}}</select>',
-            'textarea' => '<textarea name="{{name}}" class="form-control"{{attrs}}>{{value}}</textarea>',
+            //Container element user for checkboxes
+            'checkboxContainer' => '<div class="input mb-3 form-check{{required}}">{{content}}{{help}}</div>',
+            //Container element user for checkboxes when has an error
+            'checkboxContainerError' => '<div class="input mb-3 form-check{{required}}">{{content}}{{error}}{{help}}</div>',
+            //Error message wrapper elements
+            'error' => '<div class="invalid-feedback" id="{{id}}">{{content}}</div>',
+            //Container element used by `control()`
+            'inputContainer' => '<div class="input mb-3 {{type}}{{required}}">{{content}}{{help}}</div>',
+            //Container element used by `control()` when a field has an error
+            'inputContainerError' => '<div class="input mb-3 {{type}}{{required}} error">{{content}}{{error}}{{help}}</div>',
+            // Submit/reset button
+            'inputSubmit' => '<button{{attrs}}>{{text}}</button>',
         ]]);
-        $this->_defaultWidgets['hidden'] = [HiddenWidget::class];
 
         parent::__construct($view, $config);
+
+        $this->isPost = $this->getView()->getRequest()->is('post');
     }
 
     /**
-     * Creates a button.
-     *
-     * This method creates a button. To create a POST button, you should use
-     *  the `postButton()` method.
-     * Instead, to create a link with the appearance of a button, you should
-     *  use the `button()` method provided by `HtmlHelper`.
-     * @param string $title The button label or an image
-     * @param array $options HTML attributes and options
-     * @return string
-     * @see postButton()
-     * @see \MeTools\View\Helper\HtmlHelper::button()
+     * Generates an input element
+     * @param string $fieldName the field name
+     * @param array<string, mixed> $options The options for the input element
+     * @return array|string The generated input element string
+     *  or array if checkbox() is called with option 'hiddenField' set to '_split'
      */
-    public function button(string $title = '', array $options = []): string
+    protected function _getInput(string $fieldName, array $options)
+    {
+        $options = optionsParser($options);
+
+        //Class (checkboxes and radios have their own class)
+        if (!in_array($options->get('type'), ['checkbox', 'radio'])) {
+            $options->append('class', 'form-control');
+        }
+
+        /**
+         * Add class on `post` request (the form has been filled out)
+         * @see https://getbootstrap.com/docs/5.2/forms/validation/#server-side
+         */
+        if ($this->isPost) {
+            $options->append('class', $this->isFieldError($fieldName) ? 'is-invalid' : 'is-valid');
+        }
+
+        return parent::_getInput($fieldName, $options->toArray());
+    }
+
+    /**
+     * Generate label for input
+     * @param string $fieldName The name of the field to generate label for
+     * @param array<string, mixed> $options Options list
+     * @return string|false Generated label element or false
+     */
+    protected function _getLabel(string $fieldName, array $options)
+    {
+        if ($options['label'] === false) {
+            return false;
+        }
+
+        $label = optionsParser(is_string($options['label']) ? ['text' => $options['label']] : ($options['label'] ?? []));
+
+        //Checkbox and inline form fields have their own label class
+        if ($options['type'] === 'checkbox') {
+            $class = 'form-check-label';
+        } elseif ($this->isInline()) {
+            $class = 'visually-hidden';
+        }
+        $label->append('class', $class ?? 'form-label');
+
+        return parent::_getLabel($fieldName, ['label' => $label->toArray()] + $options);
+    }
+
+    /**
+     * Returns the input type that was guessed for the provided fieldName,
+     * based on the internal type it is associated too, its name and the
+     * variables that can be found in the view template
+     * @param string $fieldName the name of the field to guess a type for
+     * @param array<string, mixed> $options the options passed to the input method
+     * @return string
+     */
+    protected function _inputType(string $fieldName, array $options): string
+    {
+        //Forces the `password` type if `$fieldName` contains "password" or "pwd" words
+        if (str_contains($fieldName, 'password') || str_contains($fieldName, 'pwd')) {
+            return 'password';
+        }
+
+        return parent::_inputType($fieldName, $options);
+    }
+
+    /**
+     * Creates a `<button>` tag.
+     *
+     * See the parent method for all available options.
+     * @param string $title The button's caption. Not automatically HTML encoded
+     * @param array<string, mixed> $options Array of options and HTML attributes
+     * @return string A HTML button tag
+     */
+    public function button(string $title, array $options = []): string
     {
         $options = optionsParser($options, ['escapeTitle' => false, 'type' => 'button']);
         $options->addButtonClasses($options->contains('type', 'submit') ? 'success' : 'primary');
@@ -98,227 +164,183 @@ class FormHelper extends CakeFormHelper
     }
 
     /**
-     * Creates a checkbox element
-     * @param string $fieldName Field name, should be "Modelname.fieldname"
-     * @param array $options HTML attributes and options
-     * @return string|array<string> An HTML text input element
+     * Creates a CKEditor textarea.
+     *
+     * To add the scripts for CKEditor, you should use the `LibraryHelper`.
+     * @param string $fieldName This should be "modelname.fieldname"
+     * @param array<string, mixed> $options Each type of input takes different options
+     * @return string
+     * @see \MeTools\View\Helper\LibraryHelper::ckeditor()
+     */
+    public function ckeditor(string $fieldName, array $options = []): string
+    {
+        $options = optionsParser($options, ['label' => false, 'type' => 'textarea']);
+        $options->append('class', 'wysiwyg editor');
+
+        return $this->control($fieldName, $options->toArray());
+    }
+
+    /**
+     * Creates a checkbox input widget.
+     *
+     * See the parent method for all available options.
+     * @param string $fieldName Name of a field, like this "modelname.fieldname"
+     * @param array<string, mixed> $options Array of HTML attributes
+     * @return array<string>|string An HTML text input element
      */
     public function checkbox(string $fieldName, array $options = [])
     {
-        $options = optionsParser($options);
+        $options = optionsParser($options)->append('class', 'form-check-input');
 
-        if (!$options->exists('hiddenField') || !$options->contains('hiddenField', false)) {
-            $options->add('hiddenField', true);
+        if ($this->isInline()) {
+            $this->setTemplates([
+                'checkboxContainer' => '<div class="col-12"><div class="form-check{{required}}">{{content}}</div></div>',
+                'checkboxContainerError' => '<div class="col-12"><div class="form-check{{required}} error">{{content}}{{error}}</div></div>',
+            ]);
         }
 
         return parent::checkbox($fieldName, $options->toArray());
     }
 
     /**
-     * Creates a CKEditor textarea.
+     * Generates a form control element complete with label and wrapper div.
      *
-     * To add the scripts for CKEditor, you should use the `LibraryHelper`.
-     * @param string $fieldName Field name, should be "Modelname.fieldname"
-     * @param array $options HTML attributes and options
-     * @return string
-     * @see \MeTools\View\Helper\LibraryHelper::ckeditor()
-     */
-    public function ckeditor(string $fieldName, array $options = []): string
-    {
-        $options = optionsParser($options, ['label' => false, 'type' => 'textarea'])->append('templates', [
-            'textarea' => '<textarea name="{{name}}" class="form-control wysiwyg editor"{{attrs}}>{{value}}</textarea>',
-        ]);
-
-        return $this->control($fieldName, $options->toArray());
-    }
-
-    /**
-     * Generates an input element complete with label and wrapper div
-     * @param string $fieldName Field name, should be "Modelname.fieldname"
-     * @param array $options HTML attributes and options
-     * @return string
+     * See the parent method for all available options.
+     * @param string $fieldName This should be "modelname.fieldname"
+     * @param array<string, mixed> $options Each type of input takes different options
+     * @return string Completed form widget
      */
     public function control(string $fieldName, array $options = []): string
     {
         $this->resetTemplates();
-
         $options = optionsParser($options);
 
-        //If the name contains the "password" word, then the type is `password`
-        if (str_contains($fieldName, 'password')) {
-            $options->addDefault(['type' => 'password']);
-        }
-
-        //Gets the input type
-        $type = $options->get('type') ?: $this->_inputType($fieldName, $options->toArray());
-        if ($type === 'select' && !$options->exists('default') && !$options->exists('value')) {
-            $options->addDefault(['empty' => true]);
-        }
-
-        //Help text (form text)
-        //See https://getbootstrap.com/docs/5.2/forms/overview/#form-text
-        if ($options->exists('help')) {
-            $help = array_map(fn(string $help): string => $this->Html->div('form-text text-muted', trim($help)), (array)$options->consume('help'));
-            $options->append('templateVars', ['help' => implode('', $help)]);
-        }
-
-        //Input group. Fixes templates
-        //See https://getbootstrap.com/docs/5.2/forms/input-group
-        if ($options->exists('button')) {
-            $options->append([
-                'templates' => ['formGroup' => '{{label}}<div class="input-group">{{input}}{{button}}</div>'],
-                'templateVars' => ['button' => $this->Html->div('input-group-append', $options->consume('button'))],
+        /**
+         * Inline forms.
+         * By default, no help blocks.
+         * @see https://getbootstrap.com/docs/5.2/forms/layout/#inline-forms
+         */
+        if ($this->isInline()) {
+            $this->setTemplates([
+                'inputContainer' => '<div class="col-12 {{type}}{{required}}">{{content}}</div>',
+                'inputContainerError' => '<div class="col-12 {{type}}{{required}} error">{{content}{{error}}</div>',
             ]);
         }
 
-        //If is an inline form
-        if ($this->inline) {
-            //By default, no help blocks
-            $options->append('templates', [
-                'inputContainer' => '<div class="form-group input {{type}}{{required}}">{{content}}</div>',
+        if ($options->get('type') === 'radio') {
+            $this->setTemplates(['nestingLabel' => '<div class="form-check">{{hidden}}{{input}}<label{{attrs}}>{{text}}</label></div>']);
+        }
+
+        /**
+         * Help text (form text).
+         * These are ignored in inline forms.
+         * @see https://getbootstrap.com/docs/5.2/forms/overview/#form-text
+         */
+        if ($options->exists('help') && !$this->isInline()) {
+            $help = implode('', array_map(fn(string $help): string => $this->Html->div('form-text text-muted', trim($help)), (array)$options->consume('help')));
+            $options->append('templateVars', compact('help'));
+        }
+
+        /**
+         * Input group (`append-text` and `prepend-text` options).
+         * It can also handle buttons.
+         * @see https://getbootstrap.com/docs/5.2/forms/input-group
+         */
+        if ($options->exists('append-text') || $options->exists('prepend-text')) {
+            $this->setTemplates([
+                'formGroup' => '{{label}}<div class="input-group' . ($this->isPost ? ' has-validation' : '') . '">{{prepend}}{{input}}{{append}}{{error}}</div>',
+                'inputContainer' => '<div class="input mb-3 {{type}}{{required}}">{{content}}{{help}}</div>',
+                'inputContainerError' => '<div class="input mb-3 {{type}}{{required}} error">{{content}}{{help}}</div>',
             ]);
 
-            //If it is not a checkbox
-            if ($type !== 'checkbox' && (!$options->exists('label') || $options->get('label') !== false)) {
-                $label = $options->get('label') ? ['text' => $options->get('label')] : [];
-                $label = optionsParser($label)->append('class', 'sr-only');
-                $options->add('label', $label->toArray());
+            foreach (['append', 'prepend'] as $name) {
+                $value = $options->consume($name . '-text') ?: '';
+                if ($value && !str_starts_with($value, '<button') && !str_starts_with($value, '<div class="submit')) {
+                    $value = $this->Html->span($value, ['class' => 'input-group-text']);
+                }
+                $templateVars[$name] = $value;
             }
+            $options->append('templateVars', $templateVars);
         }
 
         return parent::control($fieldName, $options->toArray());
     }
 
     /**
-     * Returns a `<form>` element.
+     * Returns an inline HTML form element.
+     *
+     * See the parent method for all available options.
      * @param mixed $context The context for which the form is being defined.
      *   Can be a ContextInterface instance, ORM entity, ORM resultset, or an
-     *   array of metadata. You can use `null` to make a context-less form
-     * @param array $options HTML attributes and options
-     * @return string An formatted opening `<form>` tag
-     */
-    public function create($context = null, array $options = []): string
-    {
-        $options = optionsParser($options);
-
-        //It's a form inline with the `inline` option or the `form-inline` class
-        if ($options->exists('inline') || $options->contains('class', 'form-inline')) {
-            return $this->createInline($context, $options->toArray());
-        }
-
-        return parent::create($context, $options->toArray());
-    }
-
-    /**
-     * Returns an inline form element.
-     *
-     * You can also create an inline form using the `create()` method with
-     *  the `inline` option.
-     *
-     * Note that by default `createInline` doesn't display help blocks and
-     *  errors.
-     * @param mixed $context The context for which the form is being defined.
-     *   Can be a ContextInterface instance, ORM entity, ORM resultset, or an
-     *   array of metadata. You can use `null` to make a context-less form
-     * @param array $options HTML attributes and options
-     * @return string An formatted opening `<form>` tag
+     *   array of metadata. You can use `null` to make a context-less form.
+     * @param array<string, mixed> $options An array of html attributes and options
+     * @return string A formatted opening FORM tag
+     * @see https://getbootstrap.com/docs/5.2/forms/layout/#inline-forms
      */
     public function createInline($context = null, array $options = []): string
     {
-        $this->inline = true;
-        $options = optionsParser($options)->delete('inline')->append('class', 'form-inline');
+        $this->isInline = true;
+        $options = optionsParser($options)->append('class', 'row row-cols-lg-auto g-3 align-items-center');
 
         return parent::create($context, $options->toArray());
     }
 
     /**
-     * Creates a datepicker.
+     * Closes an HTML form.
      *
-     * To add the scripts for datepicker, you should use the `LibraryHelper`.
-     * @param string $fieldName Field name, should be "Modelname.fieldname"
-     * @param array $options HTML attributes and options
-     * @return string
-     * @deprecated 2.21.1 Use instead the `control()` method, which will generate a `date` input
-     */
-    public function datepicker(string $fieldName, array $options = []): string
-    {
-        deprecationWarning('Deprecated. Use instead the `control()` method, which will generate a `date` input');
-
-        return $this->control($fieldName, ['type' => 'date'] + $options);
-    }
-
-    /**
-     * Creates a datetimepicker.
-     *
-     * To add the scripts for datetimepicker, you should use the `LibraryHelper`.
-     * @param string $fieldName Field name, should be "Modelname.fieldname"
-     * @param array $options HTML attributes and options
-     * @return string
-     * @deprecated 2.21.1 Use instead the `control()` method, which will generate a `datetime-local` input
-     */
-    public function datetimepicker(string $fieldName, array $options = []): string
-    {
-        deprecationWarning('Deprecated. Use instead the `control()` method, which will generate a `datetime-local` input');
-
-        return $this->control($fieldName, ['type' => 'datetime-local'] + $options);
-    }
-
-    /**
-     * Closes an HTML form, cleans up values set by `FormHelper::create()`,
-     *  and writes hidden input fields where appropriate
-     * @param array $secureAttributes Secure attributes which will be passed
-     *  as HTML attributes into the hidden input elements generated for the
-     *  Security Component.
-     * @return string
+     * See the parent method for all available options.
+     * @param array<string, mixed> $secureAttributes Secure attributes
+     * @return string A closing FORM tag
      */
     public function end(array $secureAttributes = []): string
     {
-        $this->inline = false;
+        $this->isInline = false;
 
         return parent::end($secureAttributes);
     }
 
     /**
-     * Checks if the current opened form is an inline form
+     * Checks if the currently created form is an inline form
      * @return bool
      */
     public function isInline(): bool
     {
-        return $this->inline;
+        return $this->isInline;
     }
 
     /**
-     * Returns a formatted `<label>` element.
+     * Returns a formatted LABEL element for HTML forms.
+     *
      * Will automatically generate a `for` attribute if one is not provided.
-     * @param string $fieldName Field name, should be "Modelname.fieldname"
-     * @param string|null $text Text that will appear in the label field. If is
-     *  left undefined the text will be inflected from the fieldName
-     * @param array $options HTML attributes
-     * @return string
+     *
+     * See the parent method for all available options.
+     * @param string $fieldName This should be "modelname.fieldname"
+     * @param string|null $text Text that will appear in the label field. If
+     *   $text is left undefined the text will be inflected from the
+     *   fieldName
+     * @param array<string, mixed> $options An array of HTML attributes
+     * @return string The formatted LABEL element
      */
     public function label(string $fieldName, ?string $text = null, array $options = []): string
     {
         $options = optionsParser($options, ['escape' => false]);
-        $options->add('class', 'form-label' . ($this->isInline() ? ' sr-only' : ''));
-        [$text, $options] = $this->Icon->addIconToText($text, $options);
+
+        if (!$this->isInline()) {
+            $options->append('class', 'fw-bolder');
+            [$text, $options] = $this->Icon->addIconToText($text, $options);
+        }
 
         return parent::label($fieldName, $text, $options->toArray());
     }
 
     /**
-     * Creates a button with a surrounding form that submits via POST.
-     *
-     * This method creates a button in a form element. So don't use this
-     *  method in an already opened form.
-     *
-     * To create a normal button, you should use the `button()` method.
-     * To create a button with the appearance of a link, you should use the
-     *  `button()` method provided by the `HtmlHelper`.
-     * @param string $title Button title
-     * @param string|array|null $url Cake-relative URL or array of URL
-     *  parameters or external URL
-     * @param array $options Array of options and HTML attributes
-     * @return string Html code
+     * Creates a "POST button", which is a "POST link" with all the appearance
+     *  of a button
+     * @param string $title The content to be wrapped by <a> tags
+     * @param array|string|null $url Cake-relative URL or array of URL parameters, or
+     *   external URL (starts with http://)
+     * @param array<string, mixed> $options Array of HTML attributes
+     * @return string An `<a />` element
      */
     public function postButton(string $title = '', $url = null, array $options = []): string
     {
@@ -328,85 +350,72 @@ class FormHelper extends CakeFormHelper
     }
 
     /**
-     * Creates a link with a surrounding form that submits via POST.
+     * Creates a set of radio widgets.
      *
-     * This method creates a link in a form element. So don't use this method
-     *  in an already opened form.
-     *
-     * To create a normal link, you should use the `link()` method of the
-     *  `HtmlHelper`.
-     * @param string $title The content to be wrapped by <a> tags
-     * @param string|array|null $url Cake-relative URL or array of URL
-     *  parameters or external URL
-     * @param array $options Array of options and HTML attributes
-     * @return string
+     * See the parent method for all available options and attributes.
+     * @param string $fieldName Name of a field, like this "modelname.fieldname"
+     * @param iterable $options Radio button options array
+     * @param array<string, mixed> $attributes Array of attributes
+     * @return string Completed radio widget set
      */
-    public function postLink(string $title = '', $url = null, array $options = []): string
+    public function radio(string $fieldName, iterable $options = [], array $attributes = []): string
     {
-        $options = optionsParser($options, ['escape' => false, 'title' => $title]);
-        $options->add('title', trim(h(strip_tags($options->get('title') ?? ''))))->tooltip();
-        [$title, $options] = $this->Icon->addIconToText($title, $options);
+        $attributes = optionsParser($attributes)
+            ->append('class', 'form-check-input')
+            ->add('label', ['class' => 'form-check-label']);
 
-        return parent::postLink($title ?? '', $url, $options->toArray());
+        //Sets the `nestingLabel` templates only if it is still the default one,
+        //  therefore not already modified by other methods
+        if ($this->getTemplates('nestingLabel') == $this->_defaultConfig['templates']['nestingLabel']) {
+            $this->setTemplates(['nestingLabel' => '{{hidden}}{{input}}<label{{attrs}}>{{text}}</label>']);
+        }
+
+        $this->setTemplates(['label' => '']);
+
+        return parent::radio($fieldName, $options, $attributes->toArray());
     }
 
     /**
-     * Returns a formatted SELECT element
+     * Returns a formatted SELECT element.
+     *
+     * See the parent method for all available options and attributes.
      * @param string $fieldName Name attribute of the SELECT
-     * @param iterable $options Array of the OPTION elements
-     *  (as 'value'=>'Text' pairs) to be used in the SELECT element
-     * @param array $attributes The HTML attributes of the select element
-     * @return string
+     * @param iterable $options Array of the OPTION elements (as 'value'=>'Text' pairs) to be used in the
+     *   SELECT element
+     * @param array<string, mixed> $attributes The HTML attributes of the select element.
+     * @return string Formatted SELECT element
      */
     public function select(string $fieldName, iterable $options = [], array $attributes = []): string
     {
         $attributes = optionsParser($attributes);
-
         if (!$attributes->exists('default') && !$attributes->exists('value')) {
             $attributes->addDefault('empty', true);
         }
+        $attributes->append('class', 'form-select');
 
         return parent::select($fieldName, $options, $attributes->toArray());
     }
 
     /**
-     * Creates a submit button
-     * @param string|null $caption The label appearing on the submit button or
-     *  an image
-     * @param array $options HTML attributes and options
-     * @return string
+     * Creates a submit button element. This method will generate `<input />`
+     *  elements that can be used to submit, and reset forms by using $options.
+     *  Image submits can be created by supplying an image path for $caption.
+     *
+     * See the parent method for all available options.
+     * @param string|null $caption The label appearing on the button OR if string contains :// or the
+     *  extension .jpg, .jpe, .jpeg, .gif, .png use an image if the extension
+     *  exists, AND the first character is /, image is relative to webroot,
+     *  OR if the first character is not /, image is relative to webroot/img
+     * @param array<string, mixed> $options Array of options
+     * @return string An HTML submit button
      */
     public function submit(?string $caption = null, array $options = []): string
     {
-        return $this->button($caption ?: '', optionsParser($options)->add('type', 'submit')->toArray());
-    }
+        $options = optionsParser($options, ['escape' => false, 'type' => 'submit']);
+        $options->addButtonClasses($options->contains('type', 'submit') ? 'success' : 'primary');
+        [$text, $options] = $this->Icon->addIconToText($caption, $options);
+        $options->append('templateVars', ['text' => $text ?? __d('cake', 'Submit')]);
 
-    /**
-     * Creates a textarea widget
-     * @param string $fieldName Field name, should be "Modelname.fieldname"
-     * @param array $options HTML attributes and options
-     * @return string
-     */
-    public function textarea(string $fieldName, array $options = []): string
-    {
-        $options = optionsParser($options, ['cols' => null, 'rows' => null]);
-
-        return parent::textarea($fieldName, $options->toArray());
-    }
-
-    /**
-     * Creates a timepicker.
-     *
-     * To add the scripts for timepicker, you should use the `LibraryHelper`.
-     * @param string $fieldName Field name, should be "Modelname.fieldname"
-     * @param array $options HTML attributes and options
-     * @return string
-     * @deprecated 2.21.1 Use instead the `control()` method, which will generate a `time` input
-     */
-    public function timepicker(string $fieldName, array $options = []): string
-    {
-        deprecationWarning('Deprecated. Use instead the `control()` method, which will generate a `time` input');
-
-        return $this->control($fieldName, ['type' => 'time'] + $options);
+        return parent::submit($caption, $options->toArray());
     }
 }